Switch from hibernate ddl.auto=create-drop to migration

+++ Backend/src/main/java/com/auction/repository/ResetPasswordRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResetPasswordRepository extends JpaRepository<ResetPasswordEntity, Integer> {
+public interface ResetPasswordRepository extends JpaRepository<ResetPasswordEntity, Long> {
     ResetPasswordEntity findByCode(String code);
 
 }
